# cmd/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateHotelRequest:
    properties:
      address:
        example: Jl. MH Thamrin No. 1
        type: string
      city:
        example: Jakarta
        type: string
      description:
        example: Hotel bintang 5 di pusat Jakarta
        type: string
      image:
        example: https://example.com/hotel.jpg
        type: string
      name:
        example: Grand Hotel Jakarta
        type: string
    required:
    - address
    - city
    - description
    - name
    type: object
  handlers.CreateRoomRequest:
    properties:
      capacity:
        example: 2
        minimum: 1
        type: integer
      description:
        example: Kamar mewah dengan pemandangan kota
        type: string
      hotel_id:
        example: 60e6f3a89f48e1a8e8a8b123
        type: string
      image:
        example: https://example.com/room.jpg
        type: string
      name:
        example: Deluxe Room
        type: string
    required:
    - capacity
    - description
    - hotel_id
    - name
    type: object
...
  handlers.SpecialDateRequest:
    properties:
      date:
        description: Format YYYY-MM-DD
        example: "2025-12-25"
        type: string
      name:
        example: Hari Natal
        type: string
      point_cost:
        example: 3
        maximum: 3
        minimum: 1
        type: integer
      type:
        description: '"regular", "weekend", "holiday"'
        example: holiday
        type: string
    required:
    - date
    - point_cost
    - type
    type: object
  handlers.UpdateHotelRequest:
    properties:
      address:
        example: Updated address
        type: string
      city:
        example: Updated city
        type: string
      description:
        example: Updated description
        type: string
      image:
        example: https://example.com/new-image.jpg
        type: string
      name:
        example: New Hotel Name
        type: string
    type: object
...
  utils.APIErrorResponse:
    properties:
      error:
        example: Invalid input parameter
        type: string
      status:
        example: error
        type: string
    type: object
  utils.APISuccessResponse:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      status:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API untuk aplikasi pemesanan hotel berbasis point
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hotel Point API
  version: "1.0"
paths:
  /admin/dates/special:
    get:
      description: Get special dates for a date range (admin only)
      parameters:
      - description: From Date (YYYY-MM-DD)
        in: query
        name: from_date
        required: true
        type: string
      - description: To Date (YYYY-MM-DD)
        in: query
        name: to_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get special dates
      tags:
      - admin-dates
    post:
      consumes:
      - application/json
      description: Set a special date with custom point cost (admin only)
      parameters:
      - description: Special Date Information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SpecialDateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Set special date
      tags:
      - admin-dates
  /admin/dates/special/{id}:
    delete:
      description: Delete a special date (admin only)
      parameters:
      - description: Date Rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete special date
      tags:
      - admin-dates
  /admin/hotels:
    post:
      consumes:
      - application/json
      description: Create a new hotel (admin only)
      parameters:
      - description: Hotel Information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateHotelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new hotel
      tags:
      - admin-hotels
  /admin/hotels/{id}:
    delete:
      description: Delete a hotel (admin only)
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a hotel
      tags:
      - admin-hotels
    put:
      consumes:
      - application/json
      description: Update hotel information (admin only)
      parameters:
      - description: Hotel ID
        in: path
        name: id
        required: true
        type: string
      - description: Hotel Information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateHotelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a hotel
      tags:
      - admin-hotels
  /admin/rooms:
    post:
      consumes:
      - application/json
      description: Create a new room in a hotel (admin only)
      parameters:
      - description: Room Information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new room
      tags:
      - admin-rooms
  /admin/rooms/{id}:
    delete:
      description: Delete a room (admin only)
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a room
      tags:
      - admin-rooms
    put:
      consumes:
      - application/json
      description: Update room information (admin only)
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Room Information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APISuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a room
      tags:
      - admin-rooms
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"